Upgraded Dijkstra Code
Problem was that, when I was relaxing the nodes, I was updating the values in the priority queue, which was not right, we must use priority queue only for finding the minimum cost node. So I've declared an array named dis[node] which will save the cost of each node and another array named parent[node] which will save its parent node. So now we will only use this arrays for relaxing and finding the path.