DO NOT USE THIS CODE, THIS IS FOR CONCEPT ONLY
This is the naive version, do not use this code, because I coded this using the priority queue values. So use this only for concept only, by using this concept use an array that will store the cost and use this array for relaxing, priority queue is only for getting the cheapest cost node, do not use priority queue for relaxing. I have already coded a complete version in the dijkstra.cpp.