Updated the Backtracking Problem
The code was falling to infinite loop because, I was not able to check the fact that there was a chance to visit the same node twice, the parent is discovering it's child and child is discovering it's parent whenever the destination node was blocked. So I make sure that if the co ordinate was not visited before then only it can enter the priority queue. I just write a simple check for this, just checking the same co-ordinate in the solution grid. If solution grid contains 0 for the co-ordinate that means the point is never visited before.